+++ test/helios/find-bad-signature.php
@@ -2,9 +2,7 @@
 
 use S2lowLegacy\Class\VerifyPemCertificateFactory;
 use S2lowLegacy\Lib\PemCertificateFactory;
-use S2lowLegacy\Lib\PKCS12;
 use S2lowLegacy\Lib\SQLQuery;
-use S2lowLegacy\Lib\X509Certificate;
 use S2lowLegacy\Lib\XadesSignature;
 use S2lowLegacy\Lib\XadesSignatureParser;
 
@@ -28,8 +26,6 @@ $error_list = array();
 
 $xadesSignature = new XadesSignature(
     XMLSEC1_PATH,
-    new PKCS12(),
-    new X509Certificate(),
     EXTENDED_VALIDCA_PATH,
     new XadesSignatureParser(),
     new PemCertificateFactory(),